[refac] Auth Extension 네이밍 수정
