+++ src/main/kotlin/com/dcd/server/presentation/domain/auth/data/exetension/AuthResponseDataExtension.kt
@@ -1,7 +1,16 @@
 package com.dcd.server.presentation.domain.auth.data.exetension
 
 import com.dcd.server.core.domain.auth.dto.response.TokenResponseDto
+import com.dcd.server.presentation.domain.auth.data.response.ReissueTokenResponse
 import com.dcd.server.presentation.domain.auth.data.response.SignInResponse
+
+fun TokenResponseDto.toReissueResponse(): ReissueTokenResponse =
+    ReissueTokenResponse(
+        accessToken = this.accessToken,
+        accessTokenExp = this.accessTokenExp,
+        refreshToken = this.refreshToken,
+        refreshTokenExp = this.refreshTokenExp
+    )
 
 fun TokenResponseDto.toResponse(): SignInResponse =
     SignInResponse(
